Split out tokens
